changed test OpMode's servo code
Control the servo with the right stick now. Pressing A/B while the right stick is moving will set the max/min values to the current position of the servo. If the right stick isn't held down, then pressing A/B will set the current position to the max/min value. This is supposed to help us find the exact positions we want the servos to be set to.