Calabrated Odometry. The Vertical Left Encoder must be reversed and passed as the front left drive motor.
Added a template Autonomous called AutonomousTemplate that can be extended from. Then call defaultInit() which automatically initializes the OdometryDrivetrain and OdometryGlobalCoordinatePosition classes automatically. That can then be followed up with other initialization as normal, but it must be called before any methods on the OdometryDrivetrain and OdometryGlobalCoordinatePosition objects are called. View the BlueAutonomousNearSpinner as an example.

Signed-off-by: willm <willmoiser@gmail.com>