Jacob (#5)
* Create .deepsource.toml

* Self assignment of local variable detected JAVA-E0291

A local variable is assigned to itself.

This is essentially a noop but it may be indicative of a different problem. It may be that the variable shadows another in a parent scope, or that the variable may shadow a field of the object itself. Such code can cause confusion and subtle logic errors that are hard to catch.

* Non-null boxed types are inefficient JAVA-P1000

DESCRIPTION
Java's objects inherently add some overhead in terms of CPU and memory usage, and this overhead extends to boxed primitive types as well.

Avoid using the boxed object versions of primitives where they are not necessary.

When working with Java's primitives, treating them as objects can be useful at times (when using them as generic type arguments, for example). Boxed primitive types like Integer and Byte exist for this reason. These types combine the semantics of both objects and primitives; Arithmetic and logical operators apply to them as they do to primitives, but these object types also have methods such as equals() and hashCode() defined on them. They can also be null; raw primitive types can never be null.

Most boxed types have very similar names to primitives (just capitalize the first letter), and while they are convenient, they come with a number of object related overheads:

They occupy more space in memory (A header with a minimum of 8 bytes + the primitive value)
They force the JVM to allocate unnecessarily when boxing primitives into their object versions.
Only create variables with boxed types if you require primitives to be nullable; they have little utility elsewhere.

* Finalizers are deprecated since Java 9 JAVA-W0088

Finalizers have been deprecated since Java 9. They are unreliable at best and can cause performance problems and even race conditions in certain cases. Remove the finalizer from your code if possible

Finalizers are often more trouble than they are worth. The JVM is not guaranteed to call any object's finalizer due to the way java performs garbage collections. Additionally, objects with finalizers and their descendents are treated differently when garbage collection is performed, leading to reduced GC performance. The degradation can manifest as increased latency, or long pauses for GC.

Because of the way Java's garbage collection works, objects without finalizers are eligible for GC sooner than objects with them. Also, the fields of objects with no finalizer may be garbage collected along with the object that contains them. Using a finalizer prevents such code from benefiting from such optimizations.

* Method and field names must be dissimilar JAVA-W1006

This class's method names are the same as its field names, or differ from the field names only in terms of capitalization. This will cause confusion when reading the code, and will also decrease readability.

Rename the similar fields or methods so that their names are more readily distinguishable.

* This field and field AutonomousTemplate.slide have very similar names

This class's method names are the same as its field names, or differ from the field names only in terms of capitalization. This will cause confusion when reading the code, and will also decrease readability.

Rename the similar fields or methods so that their names are more readily distinguishable.

* Fields must not shadow other fields with the same name from super classes JAVA-W1007

This field appears to have the same name as a field in a super class.

This will prevent the superclass field from being accessed, and you will only be able to access the new field.

Rename the new field to avoid conflicts.

* Fields must not shadow other fields with the same name from super classes JAVA-W1007

This field appears to have the same name as a field in a super class.

This will prevent the superclass field from being accessed, and you will only be able to access the new field.

Rename the new field to avoid conflicts.

* Getter or setter methods must not perform operations other than reading or writing fields they are named after.

The main purpose of a getter is to allow easy access to private fields that the programmer wants to expose. Setters are, similarly, meant to allow users of a class to set the values of private fields.

When working with multiple fields that all need getters and/or setters, it is easy to make the mistake of just copy-pasting the same method, and forgetting to change the field that is returned/assigned, or forgetting to change the name of the method. Such mistakes can cost a great amount of time to find.

This issue is raised when:

A getter does not read the field it is named after.
A setter does not write to the field it is named after.

The function getImu did not return an IMU object. It has been renamed to getImuAngle