working on "fixing" things
changed the startup to set the maintainChassisDistance to false so we can actually test things without the robot going haywire. Had to comment out shutdownVision() in start as it was causing the robot to get stuck in start(). Got the arm to start moving in teleop by removing previousArticulation and just keeping track of specifically how long it would take to get back to home.