Refactor Joint
Joint is not performing right and looks convoluted
Refactor for redundant code
fix the speed calculations - they are off by a couple orders of
magnitude
speed smoothness has nothing to do with the scale of the numbers, when
using doubles. just do everything in degrees per second. smoothness is
dependent only on the looptime.
It doesn't update the calc when a new input target angle is supplied.
looks like confusion between oldTargetAngle and InterimAngle. there is
no relationship. this is wrong:
oldTargetAngle = interimAngle;
old angle is just there to know if a new target has been requested.
actually we don't even need to know this.