New update
Dear Ollie,
This update has a few changes, including adding proper coordinates for each of the terminals, junctions and cones. The 0,0 coordinate is meant to be the bottom left corner relative to the alliance position. Since each team has terminals in different locations, I have made two copies of each autonomous code to account for it.
If my autonomous code still shows problems during testing, please use my contingency plan code, which just gets the robot to drive 3 feet left (or right if the team is blue) to go to the nearest terminal