Lower Arm and drive closer for Front scoring of Yellow Pixel
Upper arm at a lower position relative to backdrop will reduce the bounce and thus increase the probability of scoring 20pts.
Since Upper Arm is at a lower position, the angle to the backdrop is such that we need to drive closer to the backdrop.

Delay before crossing field from wing side is increased to 15s to avoid hitting RoboWolves during scrimmage at CERC