April Tag Autonomous
Edited all 4 Autonomous programs to include the April Tags code (Without the camera and that other stuff). This doesn't include dropping the pixels (Those are in comments). There is one for Left, Mid, and Right. There is also one for if the robot didn't see anything (Just in case (Safety Net)).