Updated features for Autonomous and TeleOp
A preselect start for the teleop program after the autonomous was implemented. The autonomous class's values were edited to get the values to navigate the robot to the correct location around the field. A slight change to the name of the teleop program was made so that it would correspond to the class being called in the autonomous class. An old class that was not used was removed. A merging conflict was resolved.