Updated TeleOp and Auto
The auto program was updated to be more precise to move to the correct location on the field. A text file was made to remind about how to connect to the control hub device for future reference. The TeleOp program was changed to use the initialization class which was not used previously.