general changes and rework of TeleOpCode
Some unnecessary old comments were removed from the AutonomousCode class. The naming of the motor variables was changed to match the conventions of variable naming. Some more comments were added for clarification. Different types of motor modes were created, (with and without encoders). An array was created that has all of the motor velocities. The TeleOp was configured to work with setVelocity.