DRIVE IS GOOD ENOUGH FOR GAME DAY
Strafe seems to work. There is a bug that couldn't be repeated where
the wheels flipped to the other side of 180 deg.
Robot seems to drive in a diagonal and spin a little, guess is that
wheels aren't all aligned at the exact angle.
Goes straight enough.

This does have a problem with the arm... it can't extend properly.